chore(allowed-tags): add allowed tags "blog" page
Add a draft blog page which documents what tags are used on the site, as
well as what they are for and the prefix to use when creating a post
with those tags.

In general, the most relevant tag prefix should be used for the name of
the .md source file for a blog post.